FEATURE: Add `onClick` option
Adds the `onClick` callback called when clicking the LazyEmbed overlay,
after the overlay and the image have been hidden. It receives (a clone of)
the embed as its only argument.
Also makes the `data-src` attribute optional allowing using a LazyEmbed
with other elements than `iframe`s and loading the requested content
in the `onClick` callback.
Due to this change, renames the `iframe` variable (and related ones) to
a more general `embed`.
Moreover adds an `options` variable instead of using `this.options`.