BUGFIX: Fix `.browserslistrc`
Renames the `.browserlistrc` file to `.browserslistrc`, specifies the
minimum version for browserslist to be compatible with and the
browserslist configuration and switch from `gulp-cssnano` to
`gulp-postcss` and `cssnano` (version 4 RC) to use a recent version of
autoprefixer so that `postcss-merge-rules` supports the browserslist
configuration. Also uses `autoprefixer` via `gulp-postcss` instead of
`gulp-autoprefixer`.